Lazily retrieve files and comments
Right now we get the files and comments associated with a pull request right away and store them in `PullRequestInfo`. However, each of these requires calling back into the GitHub API, and in many cases we don't need the information. For example, if the base branch is unprotected we don't need to bother with the files or comments at all, and we only need the comments if the branch is protected *and* we've touched an XLIFF file.